# vtn/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: OpenADR 3 API
  description: |
    The OpenADR 3 API supports energy retailer to energy customer Demand Response programs.
    See OpenADR 3 User Guide and Defintions for detailed descriptions of usage.
    The API includes the following capabilities and operations:

    __Manage programs:__

    * Create/Update/Delete a program
    * Search programs

    __Manage events:__

    * Create/Update/Delete an event
    * Search events

    __Manage reports:__

    * Create/Update/Delete a report
    * Search reports

    __Manage subscriptions:__

    * Create/Update/Delete subscriptions to programs, events, and reports
    * Search subscriptions
    * Subscriptions allows clients to register a callback URL (webhook) to be notified
      on the change of state of a resource

    __Manage vens:__

    * Create/Update/Delete vens and ven resources
    * Search ven and ven resources

    __Manage tokens:__

    * Obtain an access token
    * This endpoint is provided as a convenience and may be neglected in a commercial implementation
  contact:
    email: frank@pajaritotech.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 3.0.1
servers:
- url: http://localhost:8081/openadr3/3.0.1
  description: base path
paths:
...
  /vens:
    get:
      tags:
      - vens
      summary: search vens
      description: |
        List all vens.
        May filter results by venName as query param.
        May filter results by targetType and targetValues as query params.
        Use skip and pagination query params to limit response size.
      operationId: search_vens
      parameters:
      - name: venName
        in: query
        description: Indicates ven objects w venName
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: targetType
        in: query
        description: "Indicates targeting type, e.g. GROUP"
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: targetValues
        in: query
        description: "List of target values, e.g. group names"
        required: false
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
      - name: skip
        in: query
        description: number of records to skip for pagination.
        required: false
        style: form
        explode: true
        schema:
          minimum: 0
          type: integer
          format: int32
      - name: limit
        in: query
        description: maximum number of records to return.
        required: false
        style: form
        explode: true
        schema:
          maximum: 50
          minimum: 0
          type: integer
          format: int32
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ven'
                x-content-type: application/json
        "400":
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
      security:
      - oAuth2ClientCredentials:
        - read_all
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.vens_controller
    post:
      tags:
      - vens
      summary: create ven
      description: Create a new ven.
      operationId: create_ven
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ven'
        required: true
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ven'
        "400":
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "409":
          description: Conflict. Implementation dependent response if ven with same
            venIdentifier already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
      security:
      - oAuth2ClientCredentials:
        - write_vens
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.vens_controller
  /vens/{venID}:
    get:
      tags:
      - vens
      summary: search vens by ID
      description: Return the ven specified by venID specified in path.
      operationId: search_ven_by_id
      parameters:
      - name: venID
        in: path
        description: object ID of ven.
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/objectID'
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ven'
        "400":
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "404":
          description: Not Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
      security:
      - oAuth2ClientCredentials:
        - read_all
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.vens_controller
    put:
      tags:
      - vens
      summary: update  ven
      description: Update the ven specified by venID specified in path.
      operationId: update_ven
      parameters:
      - name: venID
        in: path
        description: object ID of ven.
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/objectID'
      requestBody:
        description: ven item to update.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ven'
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ven'
        "400":
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "404":
          description: Not Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "409":
          description: Conflict. Implementation dependent response if program with
            the same programName exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
      security:
      - oAuth2ClientCredentials:
        - write_vens
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.vens_controller
    delete:
      tags:
      - vens
      summary: delete  ven
      description: Delete the ven specified by venID specified in path.
      operationId: delete_ven
      parameters:
      - name: venID
        in: path
        description: object ID of ven.
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/objectID'
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ven'
        "400":
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "404":
          description: Not Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
      security:
      - oAuth2ClientCredentials:
        - write_vens
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.vens_controller
  /vens/{venID}/resources:
    get:
      tags:
      - vens
      summary: search ven resources
      description: |
        List all ven resources associated with ven with specified venID.
        May filter results by resourceName as query params.
        May filter results by targetType and targetValues as query params.
        Use skip and pagination query params to limit response size.
      operationId: search_ven_resources
      parameters:
      - name: venID
        in: path
        description: Numeric ID of ven.
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/objectID'
      - name: resourceName
        in: query
        description: Indicates resource objects with resourceName
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: targetType
        in: query
        description: "Indicates targeting type, e.g. GROUP"
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: targetValues
        in: query
        description: "List of target values, e.g. group names"
        required: false
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
      - name: skip
        in: query
        description: number of records to skip for pagination.
        required: false
        style: form
        explode: true
        schema:
          minimum: 0
          type: integer
          format: int32
      - name: limit
        in: query
        description: maximum number of records to return.
        required: false
        style: form
        explode: true
        schema:
          maximum: 50
          minimum: 0
          type: integer
          format: int32
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/resource'
                x-content-type: application/json
        "400":
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "404":
          description: Not Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
      security:
      - oAuth2ClientCredentials:
        - read_all
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.vens_controller
    post:
      tags:
      - vens
      summary: create resource
      description: Create a new resource.
      operationId: create_resource
      parameters:
      - name: venID
        in: path
        description: Numeric ID of ven.
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/objectID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/resource'
        required: true
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resource'
        "400":
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "404":
          description: Not Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "409":
          description: Conflict. Implementation dependent response if resource with  same
            resourceIdentifier already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
      security:
      - oAuth2ClientCredentials:
        - write_vens
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.vens_controller
  /vens/{venID}/resources/{resourceID}:
    get:
      tags:
      - vens
      summary: search ven resources by ID
      description: Return the ven resource specified by venID and resourceID specified
        in path.
      operationId: search_ven_resource_by_id
      parameters:
      - name: venID
        in: path
        description: object ID of the associated ven.
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/objectID'
      - name: resourceID
        in: path
        description: object ID of the resource.
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/objectID'
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resource'
        "400":
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "404":
          description: Not Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
      security:
      - oAuth2ClientCredentials:
        - read_all
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.vens_controller
    put:
      tags:
      - vens
      summary: update  ven resource
      description: Update the ven resource specified by venID and resourceID specified
        in path.
      operationId: update_ven_resource
      parameters:
      - name: venID
        in: path
        description: object ID of the associated ven.
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/objectID'
      - name: resourceID
        in: path
        description: object ID of the resource.
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/objectID'
      requestBody:
        description: resource item to update.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/resource'
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resource'
        "400":
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "404":
          description: Not Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "409":
          description: Conflict. Implementation dependent response if program with
            the same programName exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
      security:
      - oAuth2ClientCredentials:
        - write_vens
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.vens_controller
    delete:
      tags:
      - vens
      summary: delete  ven resource
      description: Delete the ven resource specified by venID and resourceID specified
        in path.
      operationId: delete_ven_resource
      parameters:
      - name: venID
        in: path
        description: object ID of the associated ven.
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/objectID'
      - name: resourceID
        in: path
        description: object ID of the resource.
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/objectID'
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resource'
        "400":
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "404":
          description: Not Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/problem'
      security:
      - oAuth2ClientCredentials:
        - write_vens
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.vens_controller
...
components:
  schemas:
...
    ven:
      required:
      - venName
      type: object
      properties:
        id:
          $ref: '#/components/schemas/objectID'
        createdDateTime:
          $ref: '#/components/schemas/dateTime'
        modificationDateTime:
          $ref: '#/components/schemas/dateTime'
        objectType:
          type: string
          description: Used as discriminator.
          enum:
          - VEN
        venName:
          maxLength: 128
          minLength: 1
          type: string
          description: |
            User generated identifier, may be VEN identifier provisioned out-of-band.
            venName is expected to be unique within the scope of a VTN
          example: VEN-999
        attributes:
          type: array
          description: A list of valuesMap objects describing attributes.
          nullable: true
          items:
            $ref: '#/components/schemas/valuesMap'
        targets:
          type: array
          description: A list of valuesMap objects describing target criteria.
          nullable: true
          items:
            $ref: '#/components/schemas/valuesMap'
        resources:
          type: array
          description: A list of resource objects representing end-devices or systems.
          nullable: true
          items:
            $ref: '#/components/schemas/resource'
      description: Ven represents a client with the ven role.
      example:
        venName: VEN-999
        createdDateTime: 2023-06-15T09:30:00Z
        resources:
        - venID: null
          createdDateTime: null
          resourceName: RESOURCE-999
          attributes:
          - null
          - null
          id: null
          modificationDateTime: null
          targets:
          - null
          - null
          objectType: RESOURCE
        - venID: null
          createdDateTime: null
          resourceName: RESOURCE-999
          attributes:
          - null
          - null
          id: null
          modificationDateTime: null
          targets:
          - null
          - null
          objectType: RESOURCE
        attributes:
        - values:
          - 0.17
          type: PRICE
        - values:
          - 0.17
          type: PRICE
        id: object-999
        modificationDateTime: null
        targets:
        - null
        - null
        objectType: VEN
...
    objectID:
      maxLength: 128
      minLength: 1
      pattern: "^[a-zA-Z0-9_-]*$"
      type: string
      description: URL safe VTN assigned object ID.
      example: object-999
...
    dateTime:
      type: string
      description: datetime in ISO 8601 format
      format: date-time
      example: 2023-06-15T09:30:00Z
...
    problem:
      type: object
      properties:
        type:
          type: string
          description: |
            An absolute URI that identifies the problem type.
            When dereferenced, it SHOULD provide human-readable documentation for the problem type
            (e.g., using HTML).
          format: uri
          example: https://zalando.github.io/problem/constraint-violation
          default: about:blank
        title:
          type: string
          description: |
            A short, summary of the problem type. Written in english and readable
            for engineers (usually not suited for non technical stakeholders and
            not localized); example: Service Unavailable.
        status:
          maximum: 600
          exclusiveMaximum: true
          minimum: 100
          type: integer
          description: |
            The HTTP status code generated by the origin server for this occurrence
            of the problem.
          format: int32
          example: 503
        detail:
          type: string
          description: |
            A human readable explanation specific to this occurrence of the
            problem.
          example: Connection to database timed out
        instance:
          type: string
          description: |
            An absolute URI that identifies the specific occurrence of the problem.
            It may or may not yield further information if dereferenced.
          format: uri
      description: |
        reusable error response. From https://opensource.zalando.com/problem/schema.yaml.
...
  securitySchemes:
    oAuth2ClientCredentials:
      type: oauth2
      description: Client credential flow.
      flows:
        clientCredentials:
          tokenUrl: auth/token
          scopes:
            read_all: VENs and BL can read all resources
            write_programs: Only BL can write to programs
            write_events: Only BL can write to events
            write_reports: only VENs can write to reports
            write_subscriptions: VENs and BL can write to subscriptions
            write_vens: VENS and BL can write to vens and resources
      x-tokenInfoFunc: swagger_server.controllers.authorization_controller.check_oAuth2ClientCredentials
      x-scopeValidateFunc: swagger_server.controllers.authorization_controller.validate_scope_oAuth2ClientCredentials
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_bearerAuth